PCIDSK SDK: fix OVERFLOW_BEFORE_WIDEN warnings (CID 1243283, 1074587, 1074572, 1074566, 1074565)
git-svn-id: https://svn.osgeo.org/gdal/trunk@31894 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/frmts/pcidsk/sdk/segment/cpcidskbitmap.cpp
@@ -237,7 +237,7 @@ int CPCIDSKBitmap::ReadBlock( int block_index, void *buffer,
                               int win_xsize, int win_ysize )
 
 {
-    uint64 block_size = (block_width * block_height + 7) / 8;
+    uint64 block_size = (static_cast<uint64>(block_width) * block_height + 7) / 8;
     uint8 *wrk_buffer = (uint8 *) buffer;
 
     if( block_index < 0 || block_index >= GetBlockCount() )
@@ -281,7 +281,7 @@ int CPCIDSKBitmap::ReadBlock( int block_index, void *buffer,
         memset( buffer, 0, (size_t) block_size );
         
         short_block_size = 
-            ((height - block_index*block_height) * block_width + 7) / 8;
+            (static_cast<uint64>(height - block_index*block_height) * block_width + 7) / 8;
         
         ReadFromFile( wrk_buffer, block_size * block_index, short_block_size );
     }
@@ -314,7 +314,7 @@ int CPCIDSKBitmap::ReadBlock( int block_index, void *buffer,
 int CPCIDSKBitmap::WriteBlock( int block_index, void *buffer )
 
 {
-    uint64 block_size = (block_width * block_height) / 8;
+    uint64 block_size = (static_cast<uint64>(block_width) * block_height) / 8;
 
     if( (block_index+1) * block_height <= height )
         WriteToFile( buffer, block_size * block_index, block_size );
@@ -323,7 +323,7 @@ int CPCIDSKBitmap::WriteBlock( int block_index, void *buffer )
         uint64 short_block_size;
 
         short_block_size =
-            ((height - block_index*block_height) * block_width + 7) / 8;
+            (static_cast<uint64>(height - block_index*block_height) * block_width + 7) / 8;
 
         WriteToFile( buffer, block_size * block_index, short_block_size );
     }
